+++ src/briefy/reflex/celery.py
@@ -11,5 +11,9 @@ def main():
     app.start()
 
 
+# TODO: create a command for this
+# to purge all tasks from all queues
+# celery -A briefy.reflex -f -Q briefy_reflex_tasks,briefy_reflex,briefy_reflex_s3 purge
+
 if __name__ == '__main__':
     main()