+++ src/briefy/reflex/tasks/kinesis.py
@@ -1,34 +1,41 @@
 """Communication with kinesis service."""
 from briefy.reflex import logger
 from briefy.reflex.celery import app
+from briefy.reflex.config import GDRIVE_DELIVERY_STREAM
 from briefy.reflex.tasks import ReflexTask
+from collections import namedtuple
+from dateutil import parser
+from datetime import datetime
 
 import boto3
+import csv
 import json
-import time
+import pytz
 
 
-GDRIVE_DELIVERY_STREAM = 'gdrive_delivery_contents_live'
+FOLDER_NAMES = [
+    'originals', 'original', 'jpeg', 'original sizes', 'original size', 'original format', 'PRINT'
+]
 
 
-@app.task(base=ReflexTask)
-def put_gdrive_record(contents: dict, order: dict, stream: str=GDRIVE_DELIVERY_STREAM) -> bool:
+@app.task(base=ReflexTask, rate_limit='10/s')
+def put_gdrive_record(result: tuple, stream: str=GDRIVE_DELIVERY_STREAM) -> bool:
     """Put gdrive folder contents and orders data in a kinesis stream.
 
-    :param contents: gdrive folder contents payload
-    :param order: order payload
+    :param result: data tuple
     :param stream: kinesis stream name
     :return: True if success and False if failed
     """
+    order, contents = result
     data = {
-        'order': order,
         'contents': contents,
+        'order': order,
     }
-
+    order_id = order.get('id')
     client = boto3.client('kinesis')
     response = client.put_record(
         Data=json.dumps(data),
-        PartitionKey=order.get('uid'),
+        PartitionKey=order_id,
         StreamName=stream
     )
     success = response['ResponseMetadata']['HTTPStatusCode'] == 200
@@ -44,6 +51,7 @@ class KinesisConsumer:
         self.stream = stream
         self.update()
         self._iterators = {}
+        self._empty_shards = []
 
         # TODO: this should be persisted
         self._sequences = {}
@@ -80,8 +88,16 @@ class KinesisConsumer:
         :return: shard iterator id
         """
         shard_id = shard['ShardId']
-        shard_iterator = self._iterators.get(shard_id)
-        if not shard_iterator:
+        shard_iterator, date = self._iterators.get(shard_id, (None, None))
+        if shard_iterator and date:
+            # validate iterator datetime and
+            date = parser.parse(date)
+            now = datetime.now().astimezone(pytz.utc)
+            date_diff = (now - date).seconds
+            if date_diff > 280:
+                shard_iterator = None
+
+        if not (shard_iterator and date):
             client = self.client
             sequence_number = self.get_sequence(shard)
             response = client.get_shard_iterator(
@@ -90,19 +106,26 @@ class KinesisConsumer:
                 ShardIteratorType='AT_SEQUENCE_NUMBER',
                 StartingSequenceNumber=sequence_number,
             )
+            date = response.get('ResponseMetadata').get('HTTPHeaders').get('date')
             shard_iterator = response['ShardIterator']
+            self._iterators[shard_id] = shard_iterator, date
+
         return shard_iterator
 
     @property
     def shards(self) -> list:
         """Stream shard ids."""
-        return self.description.get('Shards', [])
+        return [
+            shard for shard in self.description.get('Shards', [])
+            if shard['ShardId'] not in self._empty_shards
+        ]
 
-    def run(self):
+    def run(self, item_callback=None):
         """Start processing all shards."""
+        self.item_callback = item_callback
         logger.info('Starting consumer. Use CTRL+C to stop.')
-        while True:
-            time.sleep(0.5)
+        while self.shards:
+            # time.sleep(0.5)
             for shard in self.shards:
                 shard_id = shard['ShardId']
                 shard_iterator = self.get_iterator(shard)
@@ -119,15 +142,137 @@ class KinesisConsumer:
 
         if len(records) == 0:
             logger.info(f'Nothing to process for shard: "{shard_id}"')
+            self._empty_shards.append(shard_id)
         else:
             for item in records:
-                logger.info(str(item))
+                if self.item_callback:
+                    data = json.loads(item.get('Data'))
+                    order = data.get('order')
+                    contents = data.get('contents')
+                    self.item_callback(order, contents)
                 self._sequences[shard_id] = item['SequenceNumber']
 
         next_shard_iterator = response['NextShardIterator']
-        self._iterators[shard_id] = next_shard_iterator
+        date = response.get('ResponseMetadata').get('HTTPHeaders').get('date')
+        self._iterators[shard_id] = next_shard_iterator, date
+
+
+TotalAssets = namedtuple('TotalAssets', ['images', 'videos', 'others'])
+
+
+def count_assets(contents, filter_folders=False) -> TotalAssets:
+    """Count the number of assets in the gdrive folder contents result.
+
+    :param contents: briefy.gdrive.api.contents result.
+    :param filter_folders: only count images on folders with specific names.
+    :return: total number of images in the folder and sub folders.
+    """
+    if filter_folders:
+        sub_folders = [
+            folder for folder in contents.get('folders', [])
+            if folder.get('name').lower().strip() in FOLDER_NAMES
+        ]
+    else:
+        sub_folders = contents.get('folders', [])
+
+    total = TotalAssets(
+        len(contents.get('images', [])) + sum(
+            [len(folder.get('images', [])) for folder in sub_folders]
+        ),
+        len(contents.get('videos', [])) + sum(
+            [len(folder.get('videos', [])) for folder in sub_folders]
+        ),
+        len(contents.get('other_files', [])) + sum(
+            [len(folder.get('other_files', [])) for folder in sub_folders]
+        )
+    )
+
+    return total
+
+
+def export_csv(data: dict, file_path: str):
+    """Export a map of orders to csv.
+
+    :param data: dict with key, value pairs to be exported, each value should be a dict also
+    :param file_path: complete file path to save the export.
+    :return:
+    """
+    fieldnames = [
+        'briefy_id', 'number_required_assets', 'number_submissions', 'total_submissions_images',
+        'total_submissions_videos', 'total_submissions_others', 'total_archive_images',
+        'total_archive_videos', 'total_archive_others', 'total_delivery_images',
+        'total_delivery_videos', 'total_delivery_others', 'submission_links', 'archive_link',
+        'delivery_link', 'order_link'
+    ]
+
+    with open(file_path, 'w') as fout:
+        writer = csv.DictWriter(fout, fieldnames)
+        writer.writeheader()
+        for key, value in data.items():
+            writer.writerow(value)
 
 
 if __name__ == '__main__':
+    # IMPORTANT: to load the data into kinesis
+    # uri = 'https://s3.eu-central-1.amazonaws.com/ms-ophelie-dev/reports/leica/finance/20171120/20171120180500-orders-all.csv'  # noQA
+    # from briefy.reflex.tasks.leica import read_all_delivery_contents
+    # read_all_delivery_contents(uri)
+
+    TOTAL_IMG_PER_ORDER = {}
+    TO_DEBUG_LINKS = {}
+    TO_DEBUG_ZERO = {}
+    TO_DEBUG_ARCHIVE = {}
+    TO_DEBUG_SUBMISSION = {}
+
+    def callback(order, contents):
+        """Compute the values for each order."""
+        order_id = order.get('slug')
+        delivery = order.get('delivery')
+        number_required_assets = order.get('number_required_assets')
+        total_delivery = count_assets(contents.get('delivery'))
+        submissions = contents.get('submissions')
+        total_submissions = [count_assets(submission) for submission in submissions]
+        total_archive = count_assets(contents.get('archive'))
+        submission_links = ','.join(
+            [str(a.get('submission_path', 'null')) for a in order.get('assignments')]
+        )
+
+        data = {
+            'total_delivery_images': total_delivery.images,
+            'total_delivery_videos': total_delivery.videos,
+            'total_delivery_others': total_delivery.others,
+            'total_archive_images': total_archive.images,
+            'total_archive_videos': total_archive.videos,
+            'total_archive_others': total_archive.others,
+            'total_submissions_images': sum([item.images for item in total_submissions]),
+            'total_submissions_videos': sum([item.videos for item in total_submissions]),
+            'total_submissions_others': sum([item.others for item in total_submissions]),
+            'number_submissions': len(submissions),
+            'briefy_id': order_id,
+            'delivery_link': delivery.get('gdrive'),
+            'archive_link': delivery.get('archive'),
+            'submission_links': submission_links,
+            'number_required_assets': number_required_assets,
+            'order_link': f'https://app.briefy.co/orders/{order.get("id")}'
+        }
+        TOTAL_IMG_PER_ORDER[order_id] = data
+
+        if not total_delivery.images:
+            TO_DEBUG_ZERO[order_id] = data
+        elif total_delivery.images > total_archive.images:
+            TO_DEBUG_ARCHIVE[order_id] = data
+        elif data['total_submissions_images'] == 0:
+            TO_DEBUG_SUBMISSION[order_id] = data
+
+        logger.info(f'Order id {order_id} processed. Number of images: {total_delivery}')
+
     c = KinesisConsumer(GDRIVE_DELIVERY_STREAM)
-    c.run()
+    c.run(callback)
+
+    export_csv(TOTAL_IMG_PER_ORDER, '/tmp/orders-image-inventory.csv')
+    export_csv(TO_DEBUG_ARCHIVE, '/tmp/orders-inventory-check-archive.csv')
+    export_csv(TO_DEBUG_SUBMISSION, '/tmp/orders-inventory-check-submission.csv')
+    export_csv(TO_DEBUG_ZERO, '/tmp/orders-inventory-zero-images.csv')
+
+    TOTAL_IMG = sum([value.get('total_delivery_images') for value in TOTAL_IMG_PER_ORDER.values()])
+    logger.info(f'Total of delivery images found: {TOTAL_IMG}')