+++ src/briefy/reflex/tasks/leica.py
@@ -1,6 +1,7 @@
 """Tasks querying information on Leica endpoints."""
 from briefy.common.utilities.interfaces import IRemoteRestEndpoint
 from briefy.reflex import config
+from briefy.reflex import logger
 from briefy.reflex.celery import app
 from briefy.reflex.tasks import ReflexTask
 from briefy.reflex.tasks.gdrive import folder_contents
@@ -8,6 +9,7 @@ from briefy.reflex.tasks.kinesis import put_gdrive_record
 from celery import chain
 from celery import group
 from csv import DictReader
+from googleapiclient.errors import HttpError
 from io import StringIO
 from zope.component import getUtility
 
@@ -15,25 +17,26 @@ import requests
 import typing as t
 
 
-@app.task(bind=True, base=ReflexTask)
-def query_orders(self, project_id: str, states: list, page: int=1) -> t.Sequence[dict]:
+@app.task(base=ReflexTask)
+def query_orders(project_id: str='', states: list=None, page: int=1) -> t.Sequence[dict]:
     """Query orders from leica endpoint.
 
-    :param self: reference to the task class instance
     :param project_id: Project ID in the Leica database
     :param states: list of states to filter orders
     :param page: page to be returned
     :return: list of orders returned from listing payload
     """
+    states = states or []
     factory = getUtility(IRemoteRestEndpoint)
     remote = factory(config.LEICA_BASE, 'orders', 'Orders')
     params = {
         'in_state': ','.join(states),
-        'project_id': project_id,
         'current_type': 'order',
         '_page': page
     }
-    result = remote.query(params, items_per_page=100)
+    if project_id:
+        params['project_id'] = project_id
+    result = remote.query(params, items_per_page=5000)
     pagination = result['pagination']
     data = result['data']
     return data, pagination
@@ -55,7 +58,7 @@ def get_order(self, order_id: str) -> dict:
 def get_filters():
     """Get orders filter."""
     return {
-        'project_id': '6f842680-749e-4898-582d-27b240c93c34',
+        'project_id': '1dafb433-9431-4295-a349-92c4ad61c59e',
         'states': ['accepted', ],
     }
 
@@ -75,18 +78,55 @@ def orders_from_csv(csv_uri: str) -> t.Sequence[dict]:
         raise RuntimeError(f'Failure to download file from: {csv_uri}.')
 
 
+def get_folder_contents(uri: str):
+    """Return gdrive contents for a given URI and ignore 404 exceptions."""
+    result = {}
+    try:
+        if uri:
+            result = folder_contents(uri, extract_id=True, subfolders=True)
+    except HttpError as error:
+        if error.resp.status != 404:
+            raise
+    return result
+
+
+@app.task(base=ReflexTask)
+def get_assets_contents(order):
+    """Get order submissions, archive and delivery folder contents."""
+    delivery_link = order.get('delivery').get('gdrive')
+    archive_link = order.get('delivery').get('archive')
+    submission_links = [
+        assignment.get('submission_path') for assignment in order.get('assignments')
+        if assignment.get('submission_path')
+    ]
+    delivery_contents = get_folder_contents(delivery_link)
+    archive_contents = get_folder_contents(archive_link)
+    submissions_contents = [
+        get_folder_contents(link) for link in submission_links
+    ]
+    contents = {
+        'delivery': delivery_contents,
+        'archive': archive_contents,
+        'submissions': submissions_contents
+    }
+    return order, contents
+
+
 @app.task(bind=True, base=ReflexTask)
 def read_all_delivery_contents(self, csv_uri: str):
     """Read all content from the gdrive delivery folder and store in aws kinesis."""
     orders = orders_from_csv(csv_uri)
     task_list = [
         chain(
-            folder_contents.s(order.get('delivery_link'), extract_id=True),
-            put_gdrive_record.s(order)
-        ) for order in orders if order.get('delivery_link')
+            get_order.s(order.get('uid')),
+            get_assets_contents.s(),
+            put_gdrive_record.s(),
+        )
+        for order in orders if order.get('delivery_link')
     ]
+    logger.info(f'Starting a group task to collect info about {len(task_list)} Orders.')
     task_group = group(task_list)
-    return task_group().join()
+    return task_group()
 
 
 def run():
@@ -94,4 +134,4 @@ def run():
     kwargs = get_filters()
     data, pagination = query_orders.delay(**kwargs).get()
     group_task = group([get_order.s(order.get('id')) for order in data])
-    return group_task().join()
+    return group_task()